Lexical Analyzer completed!!!!!!!!!! Just have to make some tests to make sure every case is being covered

+++ src/MIPS Assembler/mips_assembler.c
@@ -12,7 +12,7 @@
 
 #include "mips_assembler.h"
 
-void print_error_msg(uint16_t line, uint8_t error) 
+void print_error_msg(uint32_t line, uint8_t error) 
 {
 	switch(error){
 		default: printf("Line %d: ERROR: error code = %d\n",line, error);
